+++ app/page.jsx
@@ -1,24 +1,25 @@
 
 import { auth } from "@/auth";
 import Aside from "./components/Aside/page";
-import LogOut from "./components/icons/LogOut";
+import LogOut from "./components/icons/LogOutIcon";
 import { signOut } from "@/auth";
 import { redirect } from "next/navigation";
 import Escuela from "./components/escuela/Escuela";
 
+
 export default async function Home() {
-   
+  
 
   const session = await auth();
  
+  
   if (!session) {
    
     redirect('/auth/login')
   }
-  else{
+   else{
     redirect("/main")
-  }
-  
+  } 
 
  
 return (
@@ -28,9 +29,10 @@ return (
         <Aside 
         role = {session.user.role.name}
         />
-        <Escuela
-        token = {session.accessToken}
-        />
+       <Escuela
+       token = {session.accessToken}
+       />
+        
         <div className="bg-[url(/images/1393565.webp')] h-full w-full ">
        
         </div>
